Add files via upload
The only difficulty you might face, is how to work with an API using JavaScript? but trust me that won't be of a big problem.

Let's talk about the application you're going to build; our weather app will show the weather information that it will get from the API like this, an Icon that shows the weather status, the temperature value (18°C) in Celsius unit, the weather description (clear sky), and then the user's city and country (London, GB).

When the user clicks on the temperature value, that will convert it from Celsius to Fahrenheit.